opuy

Contiene dos indicadores relevantes de opinión publica relativos a las elecciones en Uruguay para el periodo 1989 - 2020: intención de voto y evaluación de la gestión del presidente. Los datos son agregados y provienen de la divulgación de las empresas consultoras que realizan las encuestas.

data(opuy)

Format

data.frame con 4788 filas y 14 columnas

Details

medicion

Indicador disponible. Hasta el momento solo hay dos: Intención de voto y Evaluación de la gestión del Presidente.

empresa

Nombre de la empresa consultora que realizó la medición.

tipo_eleccion

Tipo de elección: Nacional, Balotaje o Internas. Esta variable es útil para el indicador 'Intención de voto'.

anio_medicion

Anio en el que se publicó la medición.

anio_gobierno

Anio de gobierno. Esta variable es un número entero que va desde el 1 hasta el 5 indicando respectivamente el anio de gobierno (1 = primer anio de gobierno, 2 = segundo anio de gobierno...). Es particularmente útil para el indicador 'Evaluación de la gestión del Presidente'. Con esta variable es más simple realizar comparaciones entre Presidentes según anio de gobierno.

eleccion

Esta variable refiere al anio de la elección a la que está asociada la medición. Por ejemplo, se puede tener una medición de Intención de voto del anio 2017, y esta corresponde a la eleccion de 2019.

fecha

Fecha de cierre de campo declarado por la empresa consultora. En algunos casos es posible que la fecha sea la de publicación de la medición.

partido

Nombre del partido político.

sigla

Sigla del partido político.

candidato

Nombre del candidato para el caso de la Intención de Voto en elecciones Internas y en Balotaje.

presidente

Nombre del Presidente para la medición Evaluación de la gestión del Presidente.

categoria

Categoría original de la medición para el caso de Evaluación de la gestión del Presidente. Esta variable comúnmente se mide en escala de Likert (e.g. Muy Mala, Mala, Regular, Buena, Muy buena)

categoria_unificada

Se armoniza la variable 'categoria' para poder comparar entre empresas consultoras. Los encuestados que no respondieron o que respondieron que no saben se les dió el valor de 0, los que desaprueban asumen el valor de 1, los que son indiferentes el valor de 2 y con el valor de 3 son los encuestados que respondieron que aprobaban la gestión del Presidente.

valor

Valor del indicador en porcentaje.