resumen opuy

Resumen de los indicadores que contiene la base de datos por consultora. Permite consultar todos los indicadores disponibles por consultora por anio.

Ofrece un resumen por consultora e Indicador.

Metodo print de de la funcion resumen_opuy

resumen_opuy(year = allop(), polling.org = allc)

# S3 method for opuy
summary(object, ...)

# S3 method for opuy
print(x, ...)

Arguments

year

anio de medicion

polling.org

empresa consultora: "Equipos", "Cifra", "Factum", "Interconsult", "Radar" y "Opcion".

object

objeto de la clase opuy de la funcion resumen_opuy

...

parametros adicionales

x

objeto de la clase opuy de la funcion resumen_opuy

Value

data.frame que contiene los indicadores que cada consultora realizó por anio.

data.frame

Examples

resumen_opuy()
#> 
#> 
#> ---- Resumen de Indicadores por anio y consultora ---------------------------
#> 
#> 
#>           Equipos          Cifra          Factum        Interconsult 
#> ----- --------------- --------------- --------------- ---------------
#> 1989  IV (06)|   -       -   |   -       -   |   -       -   |   -   
#> 1990  IV (01)|EG (02)    -   |   -       -   |   -       -   |   -   
#> 1991  IV (02)|EG (02)    -   |   -       -   |   -       -   |   -   
#> 1992  IV (02)|EG (03)    -   |   -       -   |   -       -   |   -   
#> 1993  IV (06)|EG (06) IV (08)|   -    IV (03)|   -       -   |   -   
#> 1994  IV (11)|EG (12) IV (11)|   -    IV (12)|   -       -   |   -   
#> 1995     -   |EG (08)    -   |   -    IV (01)|   -       -   |   -   
#> 1996     -   |EG (12)    -   |   -    IV (03)|   -       -   |   -   
#> 1997  IV (01)|EG (12)    -   |   -    IV (08)|   -       -   |   -   
#> 1998  IV (01)|EG (12)    -   |   -    IV (13)|   -    IV (01)|   -   
#> 1999  IV (09)|EG (06) IV (07)|   -    IV (16)|   -    IV (07)|   -   
#> 2000     -   |EG (03)    -   |EG (02) IV (01)|EG (07)    -   |   -   
#> 2001     -   |EG (04)    -   |   -    IV (03)|EG (07)    -   |   -   
#> 2002  IV (01)|EG (04)    -   |   -    IV (03)|EG (01)    -   |   -   
#> 2003     -   |EG (04) IV (03)|   -    IV (03)|   -    IV (02)|   -   
#> 2004  IV (07)|EG (04) IV (10)|   -    IV (15)|   -    IV (09)|   -   
#> 2005     -   |EG (05)    -   |EG (02)    -   |   -    IV (01)|EG (04)
#> 2006     -   |EG (06)    -   |   -       -   |EG (02) IV (03)|EG (04)
#> 2007     -   |EG (06)    -   |   -    IV (02)|   -    IV (05)|EG (04)
#> 2008  IV (01)|EG (06) IV (03)|EG (01) IV (04)|   -    IV (07)|EG (02)
#> 2009  IV (10)|EG (05) IV (11)|   -    IV (11)|   -    IV (15)|   -   
#> 2010  IV (03)|EG (06)    -   |EG (02) IV (02)|EG (01)    -   |EG (01)
#> 2011  IV (06)|EG (02) IV (01)|EG (04) IV (04)|   -       -   |   -   
#> 2012  IV (03)|EG (04) IV (04)|EG (06) IV (04)|   -       -   |   -   
#> 2013  IV (02)|EG (04) IV (06)|EG (03) IV (05)|   -    IV (02)|   -   
#> 2014  IV (11)|EG (05) IV (10)|EG (05) IV (14)|   -    IV (08)|   -   
#> 2015  IV (03)|EG (06)    -   |EG (02)    -   |EG (03)    -   |   -   
#> 2016  IV (05)|EG (05) IV (01)|EG (03) IV (04)|EG (03) IV (02)|   -   
#> 2017  IV (06)|EG (03) IV (03)|EG (03) IV (04)|EG (01) IV (02)|   -   
#> 2018  IV (06)|EG (04) IV (04)|EG (04) IV (04)|   -       -   |   -   
#> 2019  IV (11)|EG (01) IV (08)|EG (02) IV (10)|   -       -   |   -   
#> 2020     -   |EG (07)    -   |EG (03)    -   |   -       -   |   -   
#> 2021     -   |EG (12)    -   |EG (05)    -   |EG (02)    -   |   -   
#> 2022     -   |EG (09)    -   |EG (04)    -   |   -       -   |   -   
#> 2023     -   |   -       -   |EG (01)    -   |   -       -   |   -   
#>            Radar          Opcion     
#> ----- --------------- ---------------
#> 1989     -   |   -       -   |   -   
#> 1990     -   |   -       -   |   -   
#> 1991     -   |   -       -   |   -   
#> 1992     -   |   -       -   |   -   
#> 1993     -   |   -       -   |   -   
#> 1994     -   |   -       -   |   -   
#> 1995     -   |   -       -   |   -   
#> 1996     -   |   -       -   |   -   
#> 1997     -   |   -       -   |   -   
#> 1998  IV (04)|   -       -   |   -   
#> 1999  IV (11)|   -       -   |   -   
#> 2000     -   |   -       -   |   -   
#> 2001     -   |   -       -   |   -   
#> 2002     -   |   -       -   |   -   
#> 2003  IV (04)|   -       -   |   -   
#> 2004  IV (09)|   -       -   |   -   
#> 2005     -   |EG (02)    -   |   -   
#> 2006     -   |   -       -   |   -   
#> 2007     -   |   -       -   |   -   
#> 2008  IV (03)|   -       -   |   -   
#> 2009  IV (05)|   -       -   |   -   
#> 2010     -   |   -       -   |   -   
#> 2011     -   |   -       -   |   -   
#> 2012  IV (01)|EG (01)    -   |   -   
#> 2013  IV (01)|   -       -   |   -   
#> 2014  IV (06)|   -    IV (06)|EG (02)
#> 2015     -   |   -       -   |EG (01)
#> 2016     -   |   -    IV (03)|EG (05)
#> 2017  IV (03)|   -    IV (05)|EG (04)
#> 2018  IV (06)|   -    IV (04)|EG (04)
#> 2019  IV (16)|   -    IV (11)|EG (03)
#> 2020     -   |   -       -   |EG (02)
#> 2021     -   |   -       -   |EG (04)
#> 2022     -   |   -       -   |EG (04)
#> 2023     -   |   -       -   |EG (01)
#> 
#> 
#> IV: Intencion de Voto
#> EG: Evaluacion de gestion
#> 
#> Entre parentesis se indica la cantidad de mediciones.
#> 
#> -----------------------------------------------------------------------------
resumen_opuy(1995)
#> 
#> 
#> ---- Resumen de Indicadores por anio y consultora ---------------------------
#> 
#> 
#>           Equipos          Cifra          Factum        Interconsult 
#> ----- --------------- --------------- --------------- ---------------
#> 1995     -   |EG (08)    -   |   -    IV (01)|   -       -   |   -   
#>            Radar          Opcion     
#> ----- --------------- ---------------
#> 1995     -   |   -       -   |   -   
#> 
#> 
#> IV: Intencion de Voto
#> EG: Evaluacion de gestion
#> 
#> Entre parentesis se indica la cantidad de mediciones.
#> 
#> -----------------------------------------------------------------------------
resumen_opuy(year = 1992:1998, polling.org = c("Equipos", "Cifra"))
#> 
#> 
#> ---- Resumen de Indicadores por anio y consultora ---------------------------
#> 
#> 
#>           Equipos          Cifra     
#> ----- --------------- ---------------
#> 1992  IV (02)|EG (03)    -   |   -   
#> 1993  IV (06)|EG (06) IV (08)|   -   
#> 1994  IV (11)|EG (12) IV (11)|   -   
#> 1995     -   |EG (08)    -   |   -   
#> 1996     -   |EG (12)    -   |   -   
#> 1997  IV (01)|EG (12)    -   |   -   
#> 1998  IV (01)|EG (12)    -   |   -   
#> 
#> 
#> IV: Intencion de Voto
#> EG: Evaluacion de gestion
#> 
#> Entre parentesis se indica la cantidad de mediciones.
#> 
#> -----------------------------------------------------------------------------

foo <- resumen_opuy(year = 1992:1998, polling.org = c("Equipos", "Cifra"))
print(foo)
#> 
#> 
#> ---- Resumen de Indicadores por anio y consultora ---------------------------
#> 
#> 
#>           Equipos          Cifra     
#> ----- --------------- ---------------
#> 1992  IV (02)|EG (03)    -   |   -   
#> 1993  IV (06)|EG (06) IV (08)|   -   
#> 1994  IV (11)|EG (12) IV (11)|   -   
#> 1995     -   |EG (08)    -   |   -   
#> 1996     -   |EG (12)    -   |   -   
#> 1997  IV (01)|EG (12)    -   |   -   
#> 1998  IV (01)|EG (12)    -   |   -   
#> 
#> 
#> IV: Intencion de Voto
#> EG: Evaluacion de gestion
#> 
#> Entre parentesis se indica la cantidad de mediciones.
#> 
#> -----------------------------------------------------------------------------
summary(foo)
#> # A tibble: 3 × 6
#>   Empresa Indicador                        Mediciones Periodo   Anios Ratio
#>   <chr>   <chr>                                 <int> <chr>     <dbl> <dbl>
#> 1 Equipos Evaluacion de gestion presidente         65 1992-1998     6    11
#> 2 Equipos Intencion de voto                        21 1992-1998     6     4
#> 3 Cifra   Intencion de voto                        19 1993-1994     1    19